Even more bugs
Fireworks now executes the correct sound when you are being fired to the skies. Fixed a bug where the music manager would get disabled when you get into a new scene. Went and fixed some linking issues, where unity was not able to find the sound it was supposed to play.